+++ java-projects/spring-projects/test-containers/src/main/java/com/example/App.java
@@ -15,6 +15,7 @@
  */
 package com.example;
 
+import com.example.parameter.CreateTodoParameter;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -27,8 +28,8 @@ public class App {
     }
 
     @NotNull
-    static TodoController.CreateTodoParameter[] defaultTodoItems() {
-        return new TodoController.CreateTodoParameter[]{new TodoController.CreateTodoParameter("Testcontainersのことを調べる", "* SpringBootでの利用例\n* プロジェクトで利用できないか\n* JUnit5で使えないか"),
-                new TodoController.CreateTodoParameter("エンベデッドMySQLのことを調べる", "いろいろ使えないか")};
+    static CreateTodoParameter[] defaultTodoItems() {
+        return new CreateTodoParameter[]{new CreateTodoParameter("Testcontainersのことを調べる", "* SpringBootでの利用例\n* プロジェクトで利用できないか\n* JUnit5で使えないか"),
+                new CreateTodoParameter("エンベデッドMySQLのことを調べる", "いろいろ使えないか")};
     }
 }
