+++ java-projects/spring-projects/test-containers/src/main/java/com/example/error/Handler.java
@@ -15,7 +15,6 @@
  */
 package com.example.error;
 
-import com.example.TodoNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
