+++ java-projects/spring-projects/test-containers/src/main/java/com/example/service/TodoService.java
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.example;
+package com.example.service;
 
 import com.example.entity.Todo;
+import com.example.error.TodoNotFoundException;
 import com.example.repository.TodoRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
